Create working array of obj for cards
i worked out i needed two classes. one for the data and another to create the array. This is because if i only used one class when i tried to make an intial object i had to enter values for the first card but my method is meant to assign the values